+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 
 - New Compendium Content:
   - Demon Echelons 2-4
-- Added an "Expansion" configuration to Item Grant advancements, which allows players to drop in items. (#708)
+- Added an "Additional" configuration to Item Grant advancements, which allows players to drop in items. (#708)
 - Added suggested books and licenses in the source input form. (#841)
 - In the Power Roll Dialog, when hovering a target's name and modifiers, the target's token will be highlighted as well.